requirements.txt
amqp==2.3.2
Babel==2.6.0
backcall==0.1.0
billiard==3.5.0.3
celery==4.2.0
decorator==4.3.0
Django==2.0.6
django-bootstrap4==0.0.6
django-phonenumber-field==2.0.1
django-post-office==3.0.4
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
jsonfield==2.0.2
kombu==4.2.1
parso==0.2.1
pexpect==4.6.0
phonenumbers==8.9.14
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2018.4
redis==2.10.6
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
vine==1.1.4
wcwidth==0.1.7
